add dynamics meal configuration

+++ NutriFoods-UI/Data/Store/MealsConfiguration/MealsConfigurationState.cs
@@ -25,7 +25,7 @@ public class MealConfiguration
 {
     public MealTypes MealType { get; set; } = null!;
     
-    public double Percentage { get; set; }
+    public int Percentage { get; set; }
     
     public TimeSpan? MealTime { get; set; }
